Move the python source into a subdirectory to crate a proper package
