+++ pyproject.toml
@@ -36,8 +36,4 @@ k8grep = "k8grep:main"
 
 [tool.hatch.version]
 source = "code"
-path = "__version__.py"
-
-[tool.hatch.build.targets.wheel]
-include = [ "*.py" ]
-
+path = "k8grep/__version__.py"
